src/requirements-game-report.txt
load-dotenv~=0.1.0
google-api-python-client~=2.116.0
python-dateutil~=2.8.2
requests~=2.31.0
Jinja2~=3.1.3

